requirements.txt
streamlit>=1.32.0
openai-whisper>=20231117
torch
torchaudio
google-generativeai>=0.3.0
yt-dlp>=2024.1.0
requests>=2.31.0
